+++ app/src/main/java/fr/vinetos/tranquille/Settings.java
@@ -22,8 +22,8 @@ public class Settings extends GenericSettings {
     public static final String PREF_INCOMING_CALL_NOTIFICATIONS = "incomingCallNotifications";
     public static final String PREF_BLOCK_NEGATIVE_SIA_NUMBERS = "blockNegativeSiaNumbers";
     public static final String PREF_BLOCK_HIDDEN_NUMBERS = "blockHiddenNumbers";
-    public static final String PREF_BLOCK_BLACKLISTED = "blockBlacklisted";
-    public static final String PREF_BLACKLIST_IS_NOT_EMPTY = "blacklistIsNotEmpty";
+    public static final String PREF_BLOCK_DENYLISTED = "blockDenylisted";
+    public static final String PREF_DENYLIST_IS_NOT_EMPTY = "denylistIsNotEmpty";
     public static final String PREF_USE_CONTACTS = "useContacts";
     public static final String PREF_UI_MODE = "uiMode";
     public static final String PREF_CALL_LOG_GROUPING = "callLogGrouping";
@@ -52,7 +52,7 @@ public class Settings extends GenericSettings {
     public static final String PREF_CALL_LOG_GROUPING_DAY = "day";
 
     public static final String PREF_BLOCK_IN_LIMITED_MODE_RATING = "rating";
-    public static final String PREF_BLOCK_IN_LIMITED_MODE_BLACKLIST = "blacklist";
+    public static final String PREF_BLOCK_IN_LIMITED_MODE_DENYLIST = "denylist";
 
     static final String SYS_PREFERENCES_VERSION = "__preferencesVersion";
 
@@ -124,7 +124,7 @@ public class Settings extends GenericSettings {
     }
 
     public boolean getCallBlockingEnabled() {
-        return getBlockNegativeSiaNumbers() || getBlockHiddenNumbers() || getBlacklistEnabled();
+        return getBlockNegativeSiaNumbers() || getBlockHiddenNumbers() || getDenylistEnabled();
     }
 
     public boolean getBlockNegativeSiaNumbers() {
@@ -143,24 +143,24 @@ public class Settings extends GenericSettings {
         setBoolean(PREF_BLOCK_HIDDEN_NUMBERS, block);
     }
 
-    public boolean getBlacklistEnabled() {
-        return getBlockBlacklisted() && getBlacklistIsNotEmpty();
+    public boolean getDenylistEnabled() {
+        return getBlockDenylisted() && getDenylistIsNotEmpty();
     }
 
-    public boolean getBlockBlacklisted() {
-        return getBoolean(PREF_BLOCK_BLACKLISTED, true);
+    public boolean getBlockDenylisted() {
+        return getBoolean(PREF_BLOCK_DENYLISTED, true);
     }
 
-    public void setBlockBlacklisted(boolean block) {
-        setBoolean(PREF_BLOCK_BLACKLISTED, block);
+    public void setBlockDenylisted(boolean block) {
+        setBoolean(PREF_BLOCK_DENYLISTED, block);
     }
 
-    public boolean getBlacklistIsNotEmpty() {
-        return getBoolean(PREF_BLACKLIST_IS_NOT_EMPTY);
+    public boolean getDenylistIsNotEmpty() {
+        return getBoolean(PREF_DENYLIST_IS_NOT_EMPTY);
     }
 
-    public void setBlacklistIsNotEmpty(boolean flag) {
-        setBoolean(PREF_BLACKLIST_IS_NOT_EMPTY, flag);
+    public void setDenylistIsNotEmpty(boolean flag) {
+        setBoolean(PREF_DENYLIST_IS_NOT_EMPTY, flag);
     }
 
     public boolean getUseContacts() {
@@ -223,8 +223,8 @@ public class Settings extends GenericSettings {
         return getBlockInLimitedMode().contains(PREF_BLOCK_IN_LIMITED_MODE_RATING);
     }
 
-    public boolean isBlockingBlacklistedInLimitedModeAllowed() {
-        return getBlockInLimitedMode().contains(PREF_BLOCK_IN_LIMITED_MODE_BLACKLIST);
+    public boolean isBlockingDenylistedInLimitedModeAllowed() {
+        return getBlockInLimitedMode().contains(PREF_BLOCK_IN_LIMITED_MODE_DENYLIST);
     }
 
     public Set<String> getBlockInLimitedMode() {